Force ESM resolution for postbuild
